Refactor cluster manager
